new grouping of words where nob was in kintel
The _kintel files contain those of our candidates that also had
suggestions from Kintel; the suggestions from Kintel have an initial '+'
mark (currently, if a suggestion has no frequency numbers, it was only
in kintel).

The other files should not have bokmål words that are in Kintel,
although this still needs doing for the stuff generated from sme.